VTCCA-584 (#185)
* Adding pagination to agent clients

* Pagination work in progress

* Fixing compilation issues

* Fixed compilation issues

* added ajax pagination for manage clients

* Renaming billingAuthorityReference

* Fixing test

* added non js version for paganation

* fix javascript manage clients page size

* Adding totl row count by param.

* Line length

* added page length js

* remove client properties page

+++ app/models/PropertyLink.scala
@@ -34,7 +34,7 @@ object PropertyLink {
   implicit val format = Json.format[PropertyLink]
 }
 
-case class PropertyLinkResponse(resultCount: Option[Int],
+case class PropertyLinkResponse(resultCount: Option[Long],
                                 propertyLinks: Seq[PropertyLink])
 
 object PropertyLinkResponse {
